+++ CHANGELOG.md
@@ -704,6 +704,7 @@ Server queue and tasks
 -   Remote server can run entire books not just single prompt tasks <!-- <- TODO: !!!! -->
 -   Remote server supports callbacks / pingbacks <!-- <- TODO: !!!! -->
 -   Remote server has internal task queue <!-- <- TODO: !!!! -->
+-   Remote server can be started via `ptbk start-server`
 -   _(Maybe not implementing)_ Remote server can run in `Docker` <!-- <- TODO: !!!! -->
 -   _(Maybe not implementing)_ Remote server persists its queue in `SQLite` / `.promptbook` / `Neo4j` <!-- <- TODO: !!!! -->